requirements.txt
chromadb
langchain
sentence-transformers
python-dotenv
PyMuPDF
requests
langchain-community
langchain_huggingface
groq